feat: 코스 세부 정보에 체크포인트 추가 및 러닝 화면 개선
