+++ src/apis/courses.ts
@@ -2,6 +2,7 @@ import { useAuthStore } from "../store/authState";
 import { getDataFromS3, parseJsonl } from "./common";
 import server from "./instance";
 import {
+    Checkpoint,
     CourseDetailResponse,
     CourseResponse,
     CoursesRequest,
@@ -68,6 +69,8 @@ export async function getCourse(
     try {
         const response = await server.get(`/courses/${courseId}`);
         const telemetryUrl: string | undefined = response.data?.telemetryUrl;
+        const checkpointsUrl: string | undefined =
+            response.data?.checkpointsUrl;
 
         let telemetries: Telemetry[] = [];
         if (telemetryUrl) {
@@ -78,7 +81,16 @@ export async function getCourse(
             }
         }
 
-        return { ...response.data, telemetries };
+        let checkpoints: Checkpoint[] = [];
+        if (checkpointsUrl) {
+            const text = await getDataFromS3(checkpointsUrl);
+            if (text) {
+                const parsed = await parseJsonl(text);
+                checkpoints = (parsed as unknown as Checkpoint[]) ?? [];
+            }
+        }
+
+        return { ...response.data, telemetries, checkpoints };
     } catch (error) {
         console.error(error);
         throw error;
